Create godzilla_vs_kong.py
Filming for the highly anticipated Godzilla vs. Kong movie begins. Screenwriter Adam Wingard asks you to write a program to calculate if the projected funds are needed to make the movie. The shoot will require a certain number of extras, clothing for each extra, and decor.
It is known that:
• The set for the film is worth 10% of the budget.
• For more than 150 extras, there is a 10% clothing discount.